+++ market-socket/src/main/java/university/market/config/intercepter/WebSocketInterceptor.java
@@ -1,7 +1,10 @@
-package university.market.config.interceptor;
+package university.market.config.intercepter;
+
+import static university.market.member.utils.http.HttpRequestImpl.socketToken;
 
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.stereotype.Component;
@@ -10,32 +13,22 @@ import org.springframework.web.socket.server.HandshakeInterceptor;
 
 @Slf4j
 @Component
-public class WebSocketLoggingInterceptor implements HandshakeInterceptor {
-    private static final String WS_LOG_FORMAT = """
-            WebSocket Handshake:
-                Request URI: {}
-                Authorization: {}
-                Attributes: {}
-            ====================
-            WebSocket Response:
-                Headers: {}
-            """;
+public class WebSocketInterceptor implements HandshakeInterceptor {
 
     @Override
     public boolean beforeHandshake(ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
                                    Map<String, Object> attributes) throws Exception {
-        log.info(
-                WS_LOG_FORMAT,
-                request.getURI(),
-                request.getHeaders().get("Authorization"),
-                attributes,
-                response.getHeaders()
-        );
+        String token = request.getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
+        if (token == null) {
+            return false;
+        }
+        attributes.put(socketToken, token);
         return true;
     }
 
     @Override
     public void afterHandshake(ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
                                Exception exception) {
+
     }
 }
